Add more modules
